+++ Cargo.toml
@@ -1,3 +1,4 @@
+workspace = { members = ["crates/plugin", "crates/types", "crates/value"] }
 [package]
 name = "dwarf"
 version = "0.8.1"
@@ -43,6 +44,7 @@ async-task = { version = "4.5.0", optional = true }
 ariadne = { git = "https://github.com/uberFoo/ariadne" }
 backtrace = "0.3.67"
 bincode = "1.3.3"
+bitcode = { version = "0.5.1", features = ["serde"] }
 cfg-if = "1.0.0"
 chumsky = "0.9.3"
 circular-queue = "0.2.6"
@@ -70,8 +72,9 @@ parking_lot = { version = "0.12.1", features = [
     "hardware-lock-elision",
 ] }
 pretty_env_logger = "0.5.0"
-puteketeke = { version = "0.0.4", optional = true }
+puteketeke = { version = "0.0.5", optional = true }
 ratatui = { git = "https://github.com/uberFoo/ratatui", optional = true }
+regex = "1.10.3"
 reqwest = { version = "0.11.18", features = ["blocking", "rustls-tls"] }
 rustc-hash = "1.1.0"
 rustyline = { version = "13.0.0", optional = true, features = [
@@ -84,6 +87,7 @@ sarzak = { git = "https://github.com/uberFoo/sarzak", branch = "develop", defaul
 serde = { version = "1.0.160", features = ["derive", "rc"] }
 serde_json = "1.0.96"
 slab = "0.4.9"
+smartstring = { version = "1.0.1", features = ["serde"] }
 smol = { version = "2.0.0", optional = true }
 snafu = "0.8.0"
 syntect = { version = "5.0.0", optional = true }
@@ -104,10 +108,8 @@ tui-textarea = { git = "https://github.com/uberFoo/tui-textarea", optional = tru
 tui-tree-widget = { git = "https://github.com/uberFoo/tui-rs-tree-widget", optional = true }
 unicode-segmentation = "1.10.1"
 uuid = { version = "1.6.1", features = ["v4", "serde"] }
+# value = { path = "crates/value" }
 
-smartstring = { version = "1.0.1", features = ["serde"] }
-regex = "1.10.3"
-bitcode = { version = "0.5.1", features = ["serde"] }
 
 # [patch.crates-io]
 # puteketeke = { path = "../../github/puteketeke" }
