+++ src/chacha/value.rs
@@ -28,95 +28,11 @@ use crate::{
     ChaChaError, Context, DwarfFloat, DwarfInteger, NewRef, RefType, PATH_SEP,
 };
 
-/// The type of Enumeration Field
-///
-/// There are three types of enumeration fields: Unit, Struct, and Tuple.
-///
-/// The field descriptions refer to the following enumeration, `Foo`.
-///
-/// ```ignore
-/// enum Foo {
-///     One,
-///     Bar(int),
-///     Baz { qux: string },
-/// }
-/// ```
-///
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub enum Enum<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    /// Struct Enumeration Field
-    ///
-    /// This type of field is for when it contains a struct, as `Baz` does above.
-    /// That is to say, `Foo::Baz { qux: string }`. We store the final path element
-    /// as a string, and the struct as a `RefType<UserStruct>`.
-    Struct(RefType<Struct<T>>),
-    /// Tuple Enumeration Field
-    ///
-    /// This type of field is for when it contains a tuple, as `Bar` does above.
-    /// That is to say, `Foo::Bar(int)`.
-    ///
-    /// The type is stored as the first element of the tuple, and the path/type
-    /// as a string in the second. The third element is the enum itself.
-    Tuple((RefType<ValueType>, String), RefType<TupleEnum<T>>),
-    /// Unit Enumeration Field
-    ///
-    /// This sort of enumeration is the simplest. In `Foo`, this refers to the
-    /// field `One`: `Foo::One`. The final field in the path is stored as a
-    /// string.
-    ///
-    /// The tuple is: (Type, TypeName, FieldValue)
-    Unit(RefType<ValueType>, String, String),
-}
+pub mod _enum;
+pub mod _struct;
 
-impl<T> Enum<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    pub fn get_type(&self) -> RefType<ValueType> {
-        match self {
-            Self::Unit(ty, _, _) => ty.clone(),
-            Self::Struct(ut) => s_read!(ut).get_type().clone(),
-            Self::Tuple((ty, _), _) => ty.clone(),
-        }
-    }
-}
-
-impl<T> PartialEq for Enum<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    fn eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (Self::Unit(_a, b, e), Self::Unit(_c, d, f)) => b == d && e == f,
-            (Self::Struct(a), Self::Struct(b)) => *s_read!(a) == *s_read!(b),
-            (Self::Tuple((a, _), c), Self::Tuple((b, _), d)) => {
-                *s_read!(a) == *s_read!(b) && *s_read!(c) == *s_read!(d)
-            }
-            _ => false,
-        }
-    }
-}
-
-impl<T> Eq for Enum<T> where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default
-{
-}
-
-impl<T> fmt::Display for Enum<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Self::Unit(_, t, s) => write!(f, "{t}::{s}",),
-            Self::Struct(s) => write!(f, "{}", s_read!(s)),
-            Self::Tuple((_, t), e) => write!(f, "{t}::{}", s_read!(e)),
-        }
-    }
-}
+pub use _enum::{Enum, TupleEnum};
+pub use _struct::Struct;
 
 // 🚧 This can be deleted and replaced with just a String.
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -1523,171 +1439,4 @@ impl std::cmp::PartialEq for Value {
             (_, _) => false, //Value::Error(format!("Cannot compare {} and {}", a, b)),
         }
     }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct TupleEnum<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display,
-{
-    pub(crate) variant: String,
-    pub(crate) value: RefType<T>,
-}
-
-impl<T> PartialEq for TupleEnum<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display,
-{
-    fn eq(&self, other: &Self) -> bool {
-        self.variant == other.variant && s_read!(self.value).eq(&s_read!(other.value))
-    }
-}
-
-impl<T> Eq for TupleEnum<T> where T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display {}
-
-impl<T> TupleEnum<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display,
-{
-    pub fn new<S: AsRef<str>>(variant_name: S, value: RefType<T>) -> Self {
-        Self {
-            variant: variant_name.as_ref().to_owned(),
-            value,
-        }
-    }
-
-    pub fn value(&self) -> RefType<T> {
-        self.value.clone()
-    }
-
-    pub fn variant(&self) -> &str {
-        &self.variant
-    }
-}
-
-impl<T> fmt::Display for TupleEnum<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}({})", self.variant(), s_read!(self.value))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct Struct<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    type_name: String,
-    type_: RefType<ValueType>,
-    attrs: StructAttributes<T>,
-}
-
-impl<T> PartialEq for Struct<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    fn eq(&self, other: &Self) -> bool {
-        s_read!(self.type_).eq(&s_read!(other.type_)) && self.attrs.eq(&other.attrs)
-    }
-}
-
-impl<T> Eq for Struct<T> where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default
-{
-}
-
-impl<T> Struct<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    pub fn new<S: AsRef<str>>(type_name: S, type_: &RefType<ValueType>) -> Self {
-        Self {
-            type_name: type_name.as_ref().to_owned(),
-            type_: type_.clone(),
-            attrs: StructAttributes::default(),
-        }
-    }
-
-    /// Create a field for the user type
-    ///
-    /// This is called during type definition, from a declaration in a source file.
-    pub fn define_field<S: AsRef<str>>(&mut self, name: S, value: T) {
-        self.attrs.0.insert(name.as_ref().to_owned(), value);
-    }
-
-    pub fn get_field_value<S: AsRef<str>>(&self, name: S) -> Option<&T> {
-        self.attrs.0.get(name.as_ref())
-    }
-
-    pub fn set_field_value<S: AsRef<str>>(&mut self, name: S, value: T) -> Option<T> {
-        self.attrs.0.insert(name.as_ref().to_owned(), value)
-    }
-
-    pub fn get_type(&self) -> &RefType<ValueType> {
-        &self.type_
-    }
-
-    pub fn type_name(&self) -> &str {
-        &self.type_name
-    }
-}
-
-impl<T> fmt::Display for Struct<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let mut attrs = self.attrs.0.iter().collect::<Vec<_>>();
-        attrs.sort_by(|(k1, _), (k2, _)| k1.cmp(k2));
-
-        let name = if let Some(name) = self.type_name.strip_prefix(PATH_SEP) {
-            name
-        } else {
-            &self.type_name
-        };
-
-        let mut out = f.debug_struct(name);
-        for (k, v) in attrs {
-            out.field(k, &format_args!("{v}"));
-        }
-
-        out.finish()
-    }
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct StructAttributes<T>(HashMap<String, T>)
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default;
-
-impl<T> PartialEq for StructAttributes<T>
-where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default,
-{
-    fn eq(&self, other: &Self) -> bool {
-        if self.0.len() != other.0.len() {
-            return false;
-        }
-
-        for (k, v) in self.0.iter() {
-            if !other.0.contains_key(k) {
-                return false;
-            }
-
-            let ov = other.0.get(k).unwrap();
-
-            if !v.eq(&ov) {
-                return false;
-            }
-        }
-
-        true
-    }
-}
-
-impl<T> Eq for StructAttributes<T> where
-    T: Clone + std::fmt::Debug + PartialEq + std::fmt::Display + std::default::Default
-{
 }
